refactor: video 관리하는 함수끼리 모아둠
